Move projectsAndFolders.row outlet into core app
