+++ front/app/containers/Admin/projects/all/Lists/AdminProjectList.tsx
@@ -9,7 +9,7 @@ import { isNilOrError } from 'utils/helperUtils';
 // components
 import { SortableList, SortableRow } from 'components/admin/ResourceList';
 import ProjectRow from '../../components/ProjectRow';
-import Outlet from 'components/Outlet';
+import ProjectFolderRow from '../../projectFolders/components/ProjectFolderRow';
 import { ListHeader, HeaderTitle } from '../StyledComponents';
 import Button from 'components/UI/Button';
 
@@ -104,10 +104,9 @@ const AdminProjectList = memo<Props>((_props) => {
                             publication={item}
                           />
                         )}
-                        <Outlet
-                          id="app.containers.AdminPage.projects.all.projectsAndFolders.row"
-                          publication={item}
-                        />
+                        {item.publicationType === 'folder' && (
+                          <ProjectFolderRow publication={item} />
+                        )}
                       </StyledSortableRow>
                     );
                   }
