+++ front/app/services/permissions/index.ts
@@ -5,6 +5,7 @@ import './rules/initiativePermissions';
 import './rules/commentPermissions';
 import './rules/projectPermissions';
 import './rules/campaignPermissions';
+import './rules/projectFolderPermissions';
 
 export { hasPermission };
