+++ front/app/services/permissions/rules/projectFolderPermissions.test.ts
@@ -1,4 +1,7 @@
-import { userModeratesFolder, isProjectFolderModerator } from './roles';
+import {
+  userModeratesFolder,
+  isProjectFolderModerator,
+} from './projectFolderPermissions';
 import { makeUser } from 'services/__mocks__/users';
 
 jest.mock('modules', () => ({ streamsToReset: [] }));
